Fix clippy warnings and exclude ferrules-debug
Suppress result_large_err warnings globally via lib-level allow
attribute since FerrulesError's BlockMergeError variant is
intentionally large. Remove unnecessary f64/f32 casts in metrics,
replace loop-index patterns with iterators in table parsing, add
Default impl for OCRParser, and fix OCR batch result handling to
use idiomatic match expression. Exclude ferrules-debug from
workspace (gitignored) and remove unused kmeans dependency.

WHY:
These warnings cluttered CI output and masked real issues. The
result_large_err warnings (15 instances) were false positives for
this codebase since boxing the error would add indirection without
benefit. The kmeans crate was no longer used after title detection
was refactored.

+++ ferrules-core/src/draw.rs
@@ -140,8 +140,8 @@ pub(crate) fn draw_paths(
         for segment in &path.segments {
             match segment {
                 crate::entities::Segment::Line { start, end } => {
-                    let start = (start.0 as f32, start.1 as f32);
-                    let end = (end.0 as f32, end.1 as f32);
+                    let start = (start.0, start.1);
+                    let end = (end.0, end.1);
                     imageproc::drawing::draw_line_segment_mut(
                         &mut out_img,
                         start,
